refactor(otel): reorganize imports and add log-level utilities
